back-end-app/basicpredworker/requirements.txt
pika
psycopg2-binary
numpy
sklearn
torch